Load UERRA data into climate4R grid format

Load UERRA data into climate4R grid format

loadUerra(
  file,
  var,
  lonLim = NULL,
  latLim = NULL,
  dictionary = system.file("extdata", "UERRA_dictionary.csv", package = "TraCE21kaDSR"),
  aggr.times = NULL,
  aggr.fun = mean
)

Arguments

file

A string with the file name with the UERRA data to be loaded.

var

A string with the variable name from the file to be loaded.

lonLim

An optional vector with two values specifying the longitudinal limits (min and max) to be loaded.

latLim

An optional vector with two values specifying the latitudinal limits (min and max) to be loaded.

dictionary

An optional string with the filename with the dictionary for the UERRA data. By default it loads an internal dictionary from the package.

aggr.times

An optional numeric specifying the number of cells to be aggregated.

aggr.fun

An optional function specifying the function to be used in aggregation if aggr.times was specified.

Value

TBW

Examples

# uerra.file <- "../Data/UERRA/UERRA-HARMONIE/2m_temperature/latlon/1961-90_2m_temperature.nc"
# local.var <- "tas"
# uerra = loadUerra(uerra.file, local.var)